+++ src/Core/Streaming/StreamFactoryInterface.php
@@ -8,7 +8,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Strider2038\ImgCache\Core;
+namespace Strider2038\ImgCache\Core\Streaming;
 
 use Strider2038\ImgCache\Enum\ResourceStreamModeEnum;
