+++ src/Imaging/Image/ImageFactoryInterface.php
@@ -10,7 +10,7 @@
 
 namespace Strider2038\ImgCache\Imaging\Image;
 
-use Strider2038\ImgCache\Core\StreamInterface;
+use Strider2038\ImgCache\Core\Streaming\StreamInterface;
 use Strider2038\ImgCache\Imaging\Processing\SaveOptions;
 
 /**
