+++ src/Imaging/Processing/ImageTransformerFactoryInterface.php
@@ -10,7 +10,7 @@
 
 namespace Strider2038\ImgCache\Imaging\Processing;
 
-use Strider2038\ImgCache\Core\StreamInterface;
+use Strider2038\ImgCache\Core\Streaming\StreamInterface;
 
 /**
  * @author Igor Lazarev <strider2038@rambler.ru>
