+++ src/Imaging/Storage/Driver/YandexMapStorageDriverInterface.php
@@ -11,7 +11,7 @@
 namespace Strider2038\ImgCache\Imaging\Storage\Driver;
 
 use Strider2038\ImgCache\Core\QueryParametersCollection;
-use Strider2038\ImgCache\Core\StreamInterface;
+use Strider2038\ImgCache\Core\Streaming\StreamInterface;
 
 /**
  * @author Igor Lazarev <strider2038@rambler.ru>
