+++ tests/Unit/Imaging/Processing/Imagick/ImagickTransformerFactoryTest.php
@@ -11,8 +11,8 @@
 namespace Strider2038\ImgCache\Tests\Unit\Imaging\Processing\Imagick;
 
 use Strider2038\ImgCache\Core\FileOperationsInterface;
-use Strider2038\ImgCache\Core\StreamFactoryInterface;
-use Strider2038\ImgCache\Core\StreamInterface;
+use Strider2038\ImgCache\Core\Streaming\StreamFactoryInterface;
+use Strider2038\ImgCache\Core\Streaming\StreamInterface;
 use Strider2038\ImgCache\Imaging\Processing\Imagick\ImagickTransformer;
 use Strider2038\ImgCache\Imaging\Processing\Imagick\ImagickTransformerFactory;
 use Strider2038\ImgCache\Tests\Support\FileTestCase;
