+++ tests/Unit/Imaging/Storage/Driver/YandexMapStorageDriverTest.php
@@ -18,8 +18,8 @@ use Psr\Http\Message\StreamInterface as PsrStreamInterface;
 use Psr\Log\LoggerInterface;
 use Strider2038\ImgCache\Core\QueryParameter;
 use Strider2038\ImgCache\Core\QueryParametersCollection;
-use Strider2038\ImgCache\Core\StreamFactoryInterface;
-use Strider2038\ImgCache\Core\StreamInterface;
+use Strider2038\ImgCache\Core\Streaming\StreamFactoryInterface;
+use Strider2038\ImgCache\Core\Streaming\StreamInterface;
 use Strider2038\ImgCache\Enum\HttpMethodEnum;
 use Strider2038\ImgCache\Enum\HttpStatusCodeEnum;
 use Strider2038\ImgCache\Imaging\Storage\Driver\YandexMapStorageDriver;
